Fin Arreglos

+++ src/main/java/com/gestion_academica/gestion_academica/controllers/AlumnoController.java
@@ -1,22 +1,19 @@
 package com.gestion_academica.gestion_academica.controllers;
 
+import java.util.Optional;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
-import org.springframework.lang.NonNull;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.stereotype.Controller;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.lang.NonNull;
 
 import com.gestion_academica.gestion_academica.models.Alumno;
 import com.gestion_academica.gestion_academica.models.Asignatura;
@@ -27,7 +24,13 @@ import com.gestion_academica.gestion_academica.repos.RepoAlumno;
 import com.gestion_academica.gestion_academica.repos.RepoAsignatura;
 import com.gestion_academica.gestion_academica.repos.RepoUsuario;
 
-import jakarta.transaction.Transactional;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+
+
+
 
 
 @Controller
@@ -36,12 +39,11 @@ public class AlumnoController {
     
     @Autowired
     private RepoAlumno alumnoRepositorio;
-    
+
     @Autowired
     private RepoAsignatura asignaturaRepositorio;
 
-    @Autowired
-    private RepoUsuario usuarioRepositorio;
+    @Autowired RepoUsuario usuarioRepositorio;
 
     private final PasswordEncoder encoder;
 
@@ -52,46 +54,47 @@ public class AlumnoController {
     //Lista Alumnos
 
     @GetMapping
-    public String listarAlumnos(Model modelo){
+    public String listarAlumnos(Model modelo) {
         modelo.addAttribute("alumnos", alumnoRepositorio.findAll());
         return "alumnos/alumnos";
     }
-
+    
     //Asignaturas Matriculadas Alumno
 
     @GetMapping("/matriculas/{id}")
-    public String asignaturasMatriculadas(Model modelo, @PathVariable("id") @NonNull Long id){
+    public String asignaturasMatriculadas(Model modelo, @PathVariable("id") @NonNull Long id) {
         Alumno alumnoMatriculado;
+
         if(id != 0){
             Optional <Alumno> objAlumno = alumnoRepositorio.findById(id);
 
-            if(objAlumno.isPresent()){ alumnoMatriculado = objAlumno.get(); } else {
+            if(objAlumno.isPresent()){ alumnoMatriculado = objAlumno.get(); }else{
                 modelo.addAttribute("mensaje", "El alumno consultado no existe.");
                 return "error";
             }
-        } else {
+        }else{
             Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
             String nombreAlumnoLogueado = authentication.getName();
 
             alumnoMatriculado = alumnoRepositorio.findByUsername(nombreAlumnoLogueado);
 
             if(alumnoMatriculado == null){
-                modelo.addAttribute("mensaje", "El profesor consultado no existe.");
+                modelo.addAttribute("mensaje", "No se ha encontrado un alumno relacionado con esta cuenta");
                 return "error";
             }
         }
 
         List<Asignatura> asignaturasMatriculadas = alumnoMatriculado.getAsignaturas();
 
-            modelo.addAttribute("alumno", alumnoMatriculado);
-            modelo.addAttribute("asignaturas", asignaturasMatriculadas);
-            return "alumnos/matriculas";
+        modelo.addAttribute("alumno", alumnoMatriculado);
+        modelo.addAttribute("asignaturas", asignaturasMatriculadas);
+        return "alumnos/matriculas";
     }
 
     //Profesores Imparten Alumno
 
     @GetMapping("/profesores")
-    public String profesoresAsignados(Model modelo){
+    public String profesoresAsignados(Model modelo) {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         String nombreAlumnoLogueado = authentication.getName();
 
@@ -99,8 +102,8 @@ public class AlumnoController {
         List<Asignatura> asignaturasMatriculadas = alumnoMatriculado.getAsignaturas();
         List<Profesor> profesoresAsignados = new ArrayList<>();
 
-        for(int i = 0; i < asignaturasMatriculadas.size(); i++){
-            profesoresAsignados.add(asignaturasMatriculadas.get(i).getProfesor());
+        for(Asignatura asignatura : asignaturasMatriculadas){
+            profesoresAsignados.add(asignatura.getProfesor());
         }
 
         modelo.addAttribute("alumno", alumnoMatriculado);
@@ -123,13 +126,13 @@ public class AlumnoController {
         alumno.setAuthority("ALUMNO");
         alumnoRepositorio.save(alumno);
 
-
         Usuario usuario = new Usuario();
         usuario.setUsername(alumno.getUsername());
         usuario.setPassword(alumno.getPassword());
         usuario.setAuthority(alumno.getAuthority());
         usuario.setAlumno(alumno);
         usuarioRepositorio.save(usuario);
+        
         return "redirect:/alumnos";
     }
 
@@ -138,18 +141,18 @@ public class AlumnoController {
     @GetMapping("/edit/{id}")
     public String editAlumnoForm(Model modelo, @PathVariable("id") @NonNull Long id) {
         Optional <Alumno> objAlumno = alumnoRepositorio.findById(id);
-        if(objAlumno.isPresent()) {
-            modelo.addAttribute(
-            "alumno", objAlumno.get());
+
+        if(objAlumno.isPresent()){
+            modelo.addAttribute("alumno", objAlumno.get());
             return "alumnos/edit";
-        } else {
-            modelo.addAttribute("mensaje", "El alumno consultado no existe.");
+        }else{
+            modelo.addAttribute("mensaje", "El alumno consultado no existe");
             return "error";
         }
     }
 
     @RequestMapping(value = "/edit/{id}", method = RequestMethod.POST)
-    public String editAlumno(@PathVariable("id") @NonNull Long id, @ModelAttribute("alumnoModificado") Alumno alumnoModificado) {
+    public String editAlumno(@PathVariable("id") @NonNull Long id, @ModelAttribute("alumno") Alumno alumnoModificado) {
         Optional <Alumno> objAlumno = alumnoRepositorio.findById(id);
 
         if(objAlumno.isPresent()){
@@ -178,13 +181,12 @@ public class AlumnoController {
     @GetMapping("/delete/{id}")
     public String deleteAlumnoForm(Model modelo, @PathVariable("id") @NonNull Long id) {
         Optional <Alumno> objAlumno = alumnoRepositorio.findById(id);
-        if (objAlumno.isPresent()){
-            modelo.addAttribute(
-            "alumno", objAlumno.get());
+
+        if(objAlumno.isPresent()){
+            modelo.addAttribute("alumno", objAlumno.get());
             return "alumnos/delete";
-        } else {
-            modelo.addAttribute(
-                "mensaje", "El alumno consultado no existe.");
+        }else{
+            modelo.addAttribute("mensaje", "El alumno consultado no existe");
             return "error";
         }
     }
@@ -192,9 +194,22 @@ public class AlumnoController {
     @Transactional
     @PostMapping("/delete/{id}")
     public String deleteAlumno(@PathVariable("id") @NonNull Long id) {
-        usuarioRepositorio.deleteByAlumnoId(id);
-        alumnoRepositorio.deleteById(id);
-        return "redirect:/alumnos";
+        Optional <Alumno> objAlumno = alumnoRepositorio.findById(id);
+
+        if(objAlumno.isPresent()){
+            Alumno alumnoEliminar = objAlumno.get();
+
+            for(Asignatura asignatura : alumnoEliminar.getAsignaturas()){
+                asignatura.getAlumnos().remove(alumnoEliminar);
+                asignaturaRepositorio.save(asignatura);
+            }
+
+            usuarioRepositorio.deleteByAlumnoId(id);
+            alumnoRepositorio.deleteById(id);
+            return "redirect:/alumnos";
+        }else{
+            return "error";
+        }
     }
 
     //Matricular Alumno
@@ -202,13 +217,14 @@ public class AlumnoController {
     @GetMapping("/matricular/{id}")
     public String matricularAlumnoForm(Model modelo, @PathVariable("id") @NonNull Long id) {
         Optional <Alumno> objAlumno = alumnoRepositorio.findById(id);
-        if(objAlumno.isPresent()) {
+
+        if(objAlumno.isPresent()){
             modelo.addAttribute("alumno", objAlumno.get());
             modelo.addAttribute("asignaturas", asignaturaRepositorio.findAll());
             modelo.addAttribute("matricula", new Matricula());
             return "alumnos/matricular";
-        } else {
-            modelo.addAttribute("mensaje", "El alumno consultado no existe.");
+        }else{
+            modelo.addAttribute("mensaje", "El alumno a matricular no se ha encontrado");
             return "error";
         }
     }
@@ -225,10 +241,9 @@ public class AlumnoController {
             if(asignaturaMatriculada.getAlumnos().size() < 32){
                 alumnoMatriculado.getAsignaturas().add(asignaturaMatriculada);
                 asignaturaMatriculada.getAlumnos().add(alumnoMatriculado);
-    
+
                 alumnoRepositorio.save(alumnoMatriculado);
                 asignaturaRepositorio.save(asignaturaMatriculada);
-
                 return "redirect:/alumnos";
             }else{
                 return "error";
