+++ src/main/java/com/gestion_academica/gestion_academica/controllers/AsignaturaController.java
@@ -5,18 +5,27 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.lang.NonNull;
 import org.springframework.stereotype.Controller;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+import com.gestion_academica.gestion_academica.models.Alumno;
+import com.gestion_academica.gestion_academica.models.Asignatura;
+import com.gestion_academica.gestion_academica.models.Profesor;
+import com.gestion_academica.gestion_academica.repos.RepoAlumno;
+import com.gestion_academica.gestion_academica.repos.RepoAsignatura;
+import com.gestion_academica.gestion_academica.repos.RepoProfesor;
+
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
-import com.gestion_academica.gestion_academica.models.Asignatura;
-import com.gestion_academica.gestion_academica.repos.RepoAsignatura;
 
-import jakarta.transaction.Transactional;
+
+
+
 
 @Controller
 @RequestMapping("/asignaturas")
@@ -25,10 +34,16 @@ public class AsignaturaController {
     @Autowired
     private RepoAsignatura asignaturaRepositorio;
 
+    @Autowired
+    private RepoAlumno alumnoRepositorio;
+
+    @Autowired
+    private RepoProfesor profesorRepositorio;
+
     //Lista Asignaturas
 
     @GetMapping
-    public String listarAsignaturas(Model modelo){
+    public String listarAsignaturas(Model modelo) {
         modelo.addAttribute("asignaturas", asignaturaRepositorio.findAll());
         return "asignaturas/asignaturas";
     }
@@ -47,23 +62,23 @@ public class AsignaturaController {
         return "redirect:/asignaturas";
     }
 
-    //Editar Asignatura
+    //Editar Asignaturas
 
     @GetMapping("/edit/{id}")
     public String editAsignaturaForm(Model modelo, @PathVariable("id") @NonNull Long id) {
         Optional <Asignatura> objAsignatura = asignaturaRepositorio.findById(id);
-        if(objAsignatura.isPresent()) {
-            modelo.addAttribute(
-            "asignatura", objAsignatura.get());
+
+        if(objAsignatura.isPresent()){
+            modelo.addAttribute("asignatura", objAsignatura.get());
             return "asignaturas/edit";
-        } else {
-            modelo.addAttribute("mensaje", "La asignatura consultada no existe.");
+        }else{
+            modelo.addAttribute("mensaje", "La asignatura consultada no existe");
             return "error";
         }
     }
 
     @RequestMapping(value = "/edit/{id}", method = RequestMethod.POST)
-    public String editAsignatura(@PathVariable("id") @NonNull Long id, @ModelAttribute("asignaturaModificada") Asignatura asignaturaModificada) {
+    public String editAsignatura(@PathVariable("id") @NonNull Long id, @ModelAttribute("asignatura") Asignatura asignaturaModificada) {
         Optional <Asignatura> objAsignatura = asignaturaRepositorio.findById(id);
 
         if(objAsignatura.isPresent()){
@@ -86,13 +101,12 @@ public class AsignaturaController {
     @GetMapping("/delete/{id}")
     public String deleteAsignaturaForm(Model modelo, @PathVariable("id") @NonNull Long id) {
         Optional <Asignatura> objAsignatura = asignaturaRepositorio.findById(id);
-        if (objAsignatura.isPresent()){
-            modelo.addAttribute(
-            "asignatura", objAsignatura.get());
-            return "asignaturas/delete";
-        } else {
-            modelo.addAttribute(
-                "mensaje", "La asignatura consultada no existe.");
+
+        if(objAsignatura.isPresent()){
+            modelo.addAttribute("asignatura", objAsignatura.get());
+            return "asignatura/delete";
+        }else{
+            modelo.addAttribute("mensaje", "La asignatura no ha sido encontrada");
             return "error";
         }
     }
@@ -100,7 +114,31 @@ public class AsignaturaController {
     @Transactional
     @PostMapping("/delete/{id}")
     public String deleteAsignatura(@PathVariable("id") @NonNull Long id) {
-        asignaturaRepositorio.deleteById(id);
-        return "redirect:/asignaturas";
+        Optional <Asignatura> objAsignatura = asignaturaRepositorio.findById(id);
+
+        if(objAsignatura.isPresent()){
+            Asignatura asignaturaEliminar = objAsignatura.get();
+
+            for(Alumno alumno : asignaturaEliminar.getAlumnos()){
+                alumno.getAsignaturas().remove(asignaturaEliminar);
+                alumnoRepositorio.save(alumno);
+            }
+
+            Profesor profesorImparte = asignaturaEliminar.getProfesor();
+            profesorImparte.getAsignaturas().remove(asignaturaEliminar);
+            profesorRepositorio.save(profesorImparte);
+
+            asignaturaRepositorio.deleteById(id);
+            return "redirect:/asignaturas";
+        }else{
+            return "error";
+        }
     }
+    
+    
+    
+    
+    
+    
+    
 }
